+++ src/app/inventory/selectors.ts
@@ -2,6 +2,7 @@ import { ItemHashTag } from '@destinyitemmanager/dim-api-types';
 import { destinyVersionSelector } from 'app/accounts/selectors';
 import { currentProfileSelector, settingsSelector } from 'app/dim-api/selectors';
 import { d2ManifestSelector } from 'app/manifest/selectors';
+import { universalOrnamentPlugSetHashes } from 'app/search/d2-known-values';
 import { RootState } from 'app/store/types';
 import { emptyObject, emptySet } from 'app/utils/empty';
 import { DestinyItemPlug } from 'bungie-api-ts/destiny2';
@@ -103,7 +104,8 @@ export const materialsSelector = (state: RootState) =>
   allItemsSelector(state).filter(
     (i) =>
       i.itemCategoryHashes.includes(ItemCategoryHashes.Materials) ||
-      i.itemCategoryHashes.includes(ItemCategoryHashes.ReputationTokens)
+      i.itemCategoryHashes.includes(ItemCategoryHashes.ReputationTokens) ||
+      i.hash === 3702027555 // Spoils of Conquest do not have item category hashes
   );
 
 /** The actual raw profile response from the Bungie.net profile API */
@@ -200,6 +202,41 @@ export const ownedUncollectiblePlugsSelector = createSelector(
     }
 
     return { accountWideOwned, storeSpecificOwned };
+  }
+);
+
+/** A set containing all the hashes of unlocked PlugSet items (mods, shaders, ornaments, etc) for the given character. */
+// TODO: reconcile with other owned/unlocked selectors
+export const unlockedPlugSetItemsSelector = createSelector(
+  (_state: RootState, characterId?: string) => characterId,
+  profileResponseSelector,
+  (characterId, profileResponse) => {
+    const unlockedPlugs = new Set<number>();
+    if (profileResponse?.profilePlugSets.data?.plugs) {
+      for (const plugSetHashStr in profileResponse.profilePlugSets.data.plugs) {
+        const plugSetHash = parseInt(plugSetHashStr, 10);
+        const plugs = profileResponse.profilePlugSets.data.plugs[plugSetHash];
+        for (const plugSetItem of plugs) {
+          const useCanInsert = universalOrnamentPlugSetHashes.includes(plugSetHash);
+          if (useCanInsert ? plugSetItem.canInsert : plugSetItem.enabled) {
+            unlockedPlugs.add(plugSetItem.plugItemHash);
+          }
+        }
+      }
+    }
+    if (characterId && profileResponse?.characterPlugSets.data?.[characterId]?.plugs) {
+      for (const plugSetHashStr in profileResponse.characterPlugSets.data[characterId].plugs) {
+        const plugSetHash = parseInt(plugSetHashStr, 10);
+        const plugs = profileResponse.characterPlugSets.data[characterId].plugs[plugSetHash];
+        for (const plugSetItem of plugs) {
+          const useCanInsert = universalOrnamentPlugSetHashes.includes(plugSetHash);
+          if (useCanInsert ? plugSetItem.canInsert : plugSetItem.enabled) {
+            unlockedPlugs.add(plugSetItem.plugItemHash);
+          }
+        }
+      }
+    }
+    return unlockedPlugs;
   }
 );