+++ src/app/loadout/mod-assignment-utils.ts
@@ -15,7 +15,11 @@ import { compareBy } from 'app/utils/comparators';
 import { emptyArray } from 'app/utils/empty';
 import { getModTypeTagByPlugCategoryHash, getSpecialtySocketMetadatas } from 'app/utils/item-utils';
 import { warnLog } from 'app/utils/log';
-import { getSocketByIndex, getSocketsByCategoryHash } from 'app/utils/socket-utils';
+import {
+  getSocketByIndex,
+  getSocketsByCategoryHash,
+  plugFitsIntoSocket,
+} from 'app/utils/socket-utils';
 import { DestinyEnergyType } from 'bungie-api-ts/destiny2';
 import { SocketCategoryHashes } from 'data/d2/generated-enums';
 import _ from 'lodash';
@@ -23,7 +27,7 @@ import { upgradeSpendTierToMaxEnergy } from './armor-upgrade-utils';
 import { generateModPermutations } from './mod-permutations';
 import {
   activityModPlugCategoryHashes,
-  bucketsToCategories,
+  bucketHashToPlugCategoryHash,
   getDefaultPlugHash,
   getItemEnergyType,
 } from './mod-utils';
@@ -114,7 +118,8 @@ export function fitMostMods(
     } else {
       // possible target for plugging this mod
       const targetItem = items.find(
-        (item) => plannedMod.plug.plugCategoryHash === bucketsToCategories[item.bucket.hash]
+        (item) =>
+          plannedMod.plug.plugCategoryHash === bucketHashToPlugCategoryHash[item.bucket.hash]
       );
 
       if (targetItem) {
@@ -347,7 +352,7 @@ export function pickPlugPositions(
     // TO-DO: this is naive and is going to be misleading for armor
     if (destinationSocketIndex === -1) {
       destinationSocketIndex = existingModSockets.findIndex((socket) =>
-        socket.plugSet?.plugs.some((dimPlug) => dimPlug.plugDef.hash === modToInsert.hash)
+        plugFitsIntoSocket(socket, modToInsert.hash)
       );
     }
 
@@ -741,7 +746,7 @@ function countBucketIndependentModChangesForItem(
 
   for (const mod of bucketIndependentAssignmentsForItem) {
     const socketsThatWillFitMod = getSocketsByCategoryHash(
-      item.sockets!,
+      item.sockets,
       SocketCategoryHashes.ArmorMods
     );
     if (socketsThatWillFitMod.some((socket) => socket.plugged?.plugDef.hash === mod.hash)) {