+++ src/app/settings/SettingsPage.tsx
@@ -1,4 +1,5 @@
 import { currentAccountSelector, hasD1AccountSelector } from 'app/accounts/selectors';
+import { getStores } from 'app/bungie-api/destiny2-api';
 import { settingsSelector } from 'app/dim-api/selectors';
 import ClassIcon from 'app/dim-ui/ClassIcon';
 import { StatTotalToggle } from 'app/dim-ui/CustomStatTotal';
@@ -15,6 +16,7 @@ import DimApiSettings from 'app/storage/DimApiSettings';
 import { useThunkDispatch } from 'app/store/thunk-dispatch';
 import { clearAppBadge } from 'app/utils/app-badge';
 import { errorLog } from 'app/utils/log';
+import { download } from 'app/utils/util';
 import i18next from 'i18next';
 import exampleWeaponImage from 'images/example-weapon.jpg';
 import _ from 'lodash';
@@ -142,6 +144,16 @@ export default function SettingsPage() {
     e.preventDefault();
     window.location.reload();
     return false;
+  };
+
+  const saveProfileResponse = async () => {
+    if (currentAccount) {
+      download(
+        JSON.stringify(await getStores(currentAccount), null, '\t'),
+        'profile-data.json',
+        'application/json'
+      );
+    }
   };
 
   const itemSortOrderChanged = (sortOrder: SortProperty[]) => {
@@ -491,6 +503,16 @@ export default function SettingsPage() {
           </ErrorBoundary>
 
           <Spreadsheets />
+
+          {$DIM_FLAVOR !== 'release' && currentAccount?.destinyVersion === 2 && (
+            <section id="troubleshooting">
+              <div className="setting">
+                <button type="button" className="dim-button" onClick={saveProfileResponse}>
+                  {t('Settings.ExportProfile')}
+                </button>
+              </div>
+            </section>
+          )}
         </form>
       </PageWithMenu.Contents>
     </PageWithMenu>