+++ src/app/store-stats/MaterialCounts.tsx
@@ -10,7 +10,7 @@ import styles from './MaterialCounts.m.scss';
 
 const retiredMats = [31293053, 49145143, 1305274547, 2014411539];
 const showMats = spiderMats.filter((m) => !retiredMats.includes(m));
-const goodMats = [2979281381, 4257549984, 3853748946, 4257549985];
+const goodMats = [2979281381, 4257549984, 3853748946, 4257549985, 3702027555];
 const seasonal = [1425776985, 1776857076];
 
 export function MaterialCounts() {