+++ src/app/utils/socket-utils.ts
@@ -78,9 +78,12 @@ export function getSocketByIndex(sockets: DimSockets, socketIndex: number) {
 }
 
 /** Find all sockets on the item that belong to the given category hash */
-export function getSocketsByCategoryHash(sockets: DimSockets, categoryHash: SocketCategoryHashes) {
+export function getSocketsByCategoryHash(
+  sockets: DimSockets | null,
+  categoryHash: SocketCategoryHashes
+) {
   const category = sockets?.categories.find((c) => c.category.hash === categoryHash);
-  if (!category) {
+  if (!category || !sockets) {
     return [];
   }
   return getSocketsByIndexes(sockets, category.socketIndexes);
@@ -144,4 +147,18 @@ export function getArmorExoticPerkSocket(item: DimItem): DimSocket | undefined {
  */
 export function socketContainsIntrinsicPlug(socket: DimSocket) {
   return socket.plugged?.plugDef.plug.plugCategoryHash === PlugCategoryHashes.Intrinsics;
+}
+
+/**
+ * Is this one of the plugs that could possibly fit into this socket? This does not
+ * check whether the plug is enabled or unlocked - only that it appears in the list of
+ * possible plugs.
+ */
+export function plugFitsIntoSocket(socket: DimSocket, plugHash: number) {
+  return (
+    socket.socketDefinition.singleInitialItemHash === plugHash ||
+    (socket.plugSet
+      ? socket.plugSet.plugs.some((dimPlug) => dimPlug.plugDef.hash === plugHash)
+      : socket.plugOptions.some((p) => p.plugDef.hash === plugHash))
+  );
 }