+++ src/app/vendors/VendorItems.tsx
@@ -18,17 +18,33 @@ function vendorItemIndex(item: VendorItem) {
   return item.key;
 }
 
-function itemSort(category: string) {
+// Fix for ada-1 bounties ... https://github.com/Bungie-net/api/issues/1522
+// changes their sort to match the game
+const transmogBountyOrder = [
+  1455474223, // Vanguard
+  3675595381, // Crucible
+  2259349108, // Gambit
+  4187422269, // Destination
+  171866827, // Raid/Dungeon
+];
+
+function itemSort(vendorHash: number, category: string) {
   if (category === 'category.rank_rewards_seasonal') {
     return chainComparator<VendorItem>(
       compareBy((item) => item.item?.tier),
       compareBy(vendorItemIndex)
     );
   } else if (category === 'category_bounties') {
-    return chainComparator<VendorItem>(
-      compareBy((item) => item.item?.typeName),
-      compareBy(vendorItemIndex)
-    );
+    if (vendorHash === VENDORS.ADA_TRANSMOG) {
+      return compareBy<VendorItem>(
+        (item) => item.item?.hash && transmogBountyOrder.indexOf(item.item.hash)
+      );
+    } else {
+      return chainComparator<VendorItem>(
+        compareBy((item) => item.item?.typeName),
+        compareBy(vendorItemIndex)
+      );
+    }
   } else if (category === 'category_weapon') {
     return chainComparator<VendorItem>(compareBy((item) => item.item?.itemCategoryHashes[0]));
   } else if (category.startsWith('category_tier')) {
@@ -148,7 +164,12 @@ export default function VendorItems({
                 </h3>
                 <div className={styles.vendorItems}>
                   {items
-                    .sort(itemSort(vendor.def.displayCategories[categoryIndex]?.identifier))
+                    .sort(
+                      itemSort(
+                        vendor.def.hash,
+                        vendor.def.displayCategories[categoryIndex]?.identifier
+                      )
+                    )
                     .map(
                       (item) =>
                         item.item && (